% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global_stochasticity.R
\name{add_global_stochasticity}
\alias{add_global_stochasticity}
\title{Adds global stochasticity to a cost surface}
\usage{
add_global_stochasticity(x, percent_quantile = 1)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{percent_quantile}{\code{numeric} value between 0 and 1. See details for more information}
}
\description{
Adds stochasticity to a cost surface based on a global value to ensure that the neighbourhood adjacency is random. This method is based on 'Multiple shortest paths (MSPs)' as proposed by Pinto and Keitt (2009)
}
\details{
The add_stochasticity to a cost surface is based on the method proposed by Pinto and Keitt (2009). Rather than using a static neighbourhood (for example as supplied in the neighbours function in the create_slope_cs), the neighbourhood is redefined such that the adjacency is non-deterministic and is instead determined randomly based on the threshold value.

The algorithm proceeds as follows:

1. With a percent_quantile supplied, draw a random value between the minimum value in the cost surface and the supplied percent quantile

2. Replace values in cost surface below this random value with 0. This ensures that the conductance between the neighbours are 0, and thus deemed non-adjacent.

Supplying a percent_quantile of 0 is equivalent to incorporating no stochasticity into the cost surface. That is, if the supplied percent_quantile is 0, then no values are below this value and thus no values will be replaced with 0 (see step 2). This therefore does not change the neigbourhood adjacency.

The closer the percent_quantile is to 0, the less stochasticity is incorporated. For example, a percent_quantile value of 0.2 will result in the threshold being a random value between the minimum value in the cost surface and the 0.2 percent quantile of the values in the cost surface. All values in the cost surface below the random value will be replaced with 0 (i.e. the neighbours are no longer adjacent). In contrast, a percent_quantile value of 0.8 will result in the threshold being a random value between the minimum value in the cost surface and the 0.8 percent quantile of the values in the cost surface. In this case, there is greater probability that the random value will result in an increased number of values in the cost surface being replaced with 0.
}
\examples{

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

slope_cs2 <- add_global_stochasticity(slope_cs, percent_quantile = 0.2)
}
\references{
Pinto, N., & Keitt, T. H. (2009). Beyond the least-cost path: evaluating corridor redundancy using a graph-theoretic approach. Landscape Ecology, 24(2), 253-266. https://doi.org/10.1007/s10980-008-9303-y
}
\author{
Joseph Lewis
}
