% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_stochasticity.R
\name{add_local_stochasticity}
\alias{add_local_stochasticity}
\title{Adds local stochasticity to a cost surface}
\usage{
add_local_stochasticity(x)
}
\arguments{
\item{x}{\code{conductanceMatrix}}
}
\description{
Adds stochasticity to a cost surface based on sampling adjacent neighbours. Probability of remaining adjacent and thus traversable is equal to the probability weight of the adjacent values. Sampling done with replacement so possible for multiple adjacent neighbours to remain traversable
}
\examples{

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

slope_cs2 <- add_local_stochasticity(slope_cs)
}
\author{
Joseph Lewis
}
