% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_validation.R
\name{buffer_validation}
\alias{buffer_validation}
\title{Calculate the similarity of a least-cost path to a known route}
\usage{
buffer_validation(lcp, comparison, dist)
}
\arguments{
\item{lcp}{\code{sf}}

\item{comparison}{\code{sf}}

\item{dist}{\code{numeric} buffer distances to assess similarity}
}
\value{
\code{data.frame}
}
\description{
Calculates the similarity of a least-cost path to a known route using the buffer method proposed by Goodchild and Hunter (1997)
}
\author{
Joseph Lewis
}
