% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{calculate distance between adjacent cells}
\usage{
calculate_distance(x, adj)
}
\arguments{
\item{x}{\code{spatRaster}}

\item{adj}{\code{matrix} of adjacent cells}
}
\value{
\code{matrix} euclidean distances between adjacent cells
}
\description{
calculate distance between adjacent cells
}
\details{
calculate_distance function allows for both projected and geographic coordinate systems. If the coordinate system is geographic (e.g. wgs84) the distance is calculated using the sf::st_distance function else distance calculated using Pythagorean theorem
}
\author{
Joseph Lewis
}
