% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rmse.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculates the RMSE of a Digial Elevation Model after aggregation}
\usage{
calculate_rmse(x, y, x_rmse = 0)
}
\arguments{
\item{x}{\code{SpatRaster} original Digital Elevation Model}

\item{y}{\code{SpatRaster} aggregated Digital Elevation Model}

\item{x_rmse}{\code{numeric} RMSE of original Digital Elevation Model before aggregation}
}
\description{
Calculates the RMSE of a Digial Elevation Model after aggregation
}
\examples{

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

r2 <- terra::aggregate(r, 10)

calculate_rmse(x = r, y = r2, x_rmse = 0)
}
\author{
Joseph Lewis
}
