% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_slope_variance.R
\name{calculate_slope_variance}
\alias{calculate_slope_variance}
\title{Calculates mean standard deviation of slope values}
\usage{
calculate_slope_variance(x, neighbours, origin, destination)
}
\arguments{
\item{x}{\code{SpatRaster} Digital Elevation Model (DEM)}

\item{neighbours}{\code{numeric} value. Number of directions used in the conductance matrix calculation. Expected numeric values are 4, 8, 16, 32, 48, or matrix object. 16 (default)}

\item{origin}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}

\item{destination}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}
}
\value{
\code{numeric} value of the mean standard deviation of slope values that coincide with the straight line from the origin to the destination
}
\description{
Calculates mean standard deviation of slope values from cells that coincide with the straight line from the origin to the destination.
}
\examples{

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

locs <- sf::st_sf(geometry = sf::st_sfc(
sf::st_point(c(839769, 4199443)),
sf::st_point(c(1038608, 4100024)),
crs = terra::crs(r)))

calculate_slope_variance(x = r, neighbours = 4, origin = locs[1,], destination = locs[2,])
}
\author{
Joseph Lewis
}
