% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_locations.R
\name{check_duplicate_locations}
\alias{check_duplicate_locations}
\title{check for duplicate locations in terms of whether they share cells from a parent SpatRaster}
\usage{
check_duplicate_locations(locations_data, locations_name, cs_rast)
}
\arguments{
\item{locations_data}{\code{class("sf")} or \code{class('SpatVector')}}

\item{cs_rast}{\code{spatRaster} The DEM used in the cost-surface.}
}
\value{
\code{integer} of unique locations, expressed as cells
}
\description{
check for duplicate locations in terms of whether they share cells from a parent SpatRaster
}
\author{
Barney Harris
}
