% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locations.R
\name{check_locations}
\alias{check_locations}
\title{check supplied locations}
\usage{
check_locations(x, locations)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{locations}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}
}
\value{
\code{message}
}
\description{
checks that locations can be reached when calculating least-cost paths
}
\details{
Using the supplied conductanceMatrix, the function checks whether:
(1) the supplied locations are traversable from at least one adjacent cell
(2) the supplied locations are within the extent of the supplied conductanceMatrix
}
\author{
Joseph Lewis
}
