% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_barrier_cs.R
\name{create_barrier_cs}
\alias{create_barrier_cs}
\title{create_barrier_cs}
\usage{
create_barrier_cs(raster, barrier, neighbours = 16)
}
\arguments{
\item{raster}{\code{RasterLayer} (raster package). This is used to derive the resolution, extent, and spatial reference system to be used when calculating the barrier cost surface}

\item{barrier}{\code{Spatial*}. Areas representing barriers to movement in the landscape}

\item{neighbours}{Number of directions used in the Least Cost Path calculation. See Huber and Church (1985) for methodological considerations when choosing number of neighbours. Expected values are 4, 8, or 16. Default is 16.}
}
\value{
\code{TransitionLayer} (gdistance package) numerically expressing the barriers to movement in the landscape. The resultant \code{TransitionLayer} can be incorporated with other \code{TransitionLayer} through Raster calculations.
}
\description{
Creates a Barrier Cost Surface
}
\details{
Creates a cost surface representing barriers to movement in the landscape.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
loc1 = cbind(2667670, 6479000)
loc1 = sp::SpatialPoints(loc1)

barrier <- create_barrier_cs(raster = r, barrier = loc1)
}
\author{
Joseph Lewis
}
