% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cumulative_lcps.R
\name{create_cumulative_lcps}
\alias{create_cumulative_lcps}
\title{create_cumulative_lcps}
\usage{
create_cumulative_lcps(lcps, raster, rescale = TRUE)
}
\arguments{
\item{lcps}{Least Cost Paths. Expects Object of class SpatialLines or SpatialLinesDataFrame}

\item{raster}{Expects Object of class RasterLayer. This is used to derive extent and spatial reference system of the cumulative cost path raster.}

\item{rescale}{Rescales cost corridor raster to values between 0 and 1. Default is TRUE}
}
\description{
Creates a cumulative cost path raster using previously created LCP network
}
\details{
The create_cumulative_lcps function cumulatively combines LCPs to show LCP density and thus routes of preferential movement within the landscape.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

locs <- sp::spsample(as(r, 'SpatialPolygons'),n=10,'regular')

lcp_network <- create_lcp_network(cost_surface = final_cost_cs, locations = locs,
cost_distance = FALSE, parallel = FALSE)

cumulative_lcps <- create_cumulative_lcps(lcps = lcp_network, raster = r, rescale = FALSE)
}
\author{
Joseph Lewis
}
