% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature.R
\name{create_feature}
\alias{create_feature}
\title{create_feature}
\usage{
create_feature(raster, locations, x)
}
\arguments{
\item{raster}{Raster file. Expects Object of class RasterLayer. This is used to derive extent and spatial reference system of the landscape feature cost surface.}

\item{locations}{Locations of landscape features. Expects Object of class SpatialPoints}

\item{x}{vector of values denoting the attraction/repulsion from the landscape feature(s).}
}
\description{
Creates Landscape Feature Cost Surface to be used in Least Cost Path calculation.
}
\details{
The create_feature function computes a cost surface representating the attraction/repulsion of a feature in the landscape. The function requires a Raster (class 'RasterLayer'), the locaion of the feature(s) (class 'SpatialPoints'), and a vector (Class 'numeric') of attraction/repulsion values from the feature(s).
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
loc1 = cbind(2667670, 6479000)
loc1 = sp::SpatialPoints(loc1)

num <- seq(200, 1, length.out = 20)

feature_attraction <- create_feature(raster = r, locations = loc1, x = num)
}
\author{
Joseph Lewis
}
