% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_feature_attraction.R
\name{create_feature_attraction}
\alias{create_feature_attraction}
\title{create_feature_attraction}
\usage{
create_feature_attraction(dem, locs, max_attraction = 5,
  distance = 500, eq = "linear")
}
\arguments{
\item{dem}{Digital Elevation Model. Expects Object of class RasterLayer}

\item{locs}{Locations of landscape features. Expects Object of class SpatialPoints}

\item{max_attraction}{Value denoting the maximum attraction value of the landscape feature(s). Default is 5. Value must be above 0.}

\item{distance}{Maximum distance that the landscape feature has influence. Default is 500. Value must be above 0.}

\item{eq}{Attraction decay rate: 'linear' or 'exponential'}
}
\description{
Creates landscape feature attraction Cost Surface to be used in Least Cost Path calculation.
}
\details{
The create_feature_attraction function computes a cost surface representating the landscape feature attraction.It implements ths method proposed by Llobera (200) and uses a linear decay rate to decrease the attraction of the feature as distance increases from the feature.The function requires a Digital Elevation Model (class 'RasterLayer'),the location of the landscape(s) (class 'SpatialPoints'),the maximum attraction value of the landscape feature(s), and the maximum distance that the landscape feature has influence.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
loc1 = cbind(2667670, 6479000)
loc1 = sp::SpatialPoints(loc1)

loc2 = cbind(2667800, 6479400)
loc2 = sp::SpatialPoints(loc2)

locs <- rbind(loc1, loc2)
feature_attraction <- create_feature_attraction(dem = r, locs = locs,
max_attraction = 5, distance = 500, eq = 'linear')
}
\author{
Joseph Lewis
}
