% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_FETE_lcps.R
\name{create_FETE_lcps}
\alias{create_FETE_lcps}
\title{Calculate Least-cost Paths from each location to all other locations}
\usage{
create_FETE_lcps(x, locations, cost_distance = FALSE)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{locations}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}

\item{cost_distance}{\code{logical} if TRUE computes total accumulated cost from origin to destination. FALSE (default)}
}
\value{
\code{sf}  Least-cost paths from-everywhere-to-everywhere based on the supplied \code{conductanceMatrix}
}
\description{
Calculates Least-cost paths from-everywhere-to-everywhere. This is based on the approach proposed by White and Barber (2012).
}
\author{
Joseph Lewis
}
