% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp_network.R
\name{create_lcp_network}
\alias{create_lcp_network}
\title{create_lcp_network}
\usage{
create_lcp_network(cost_surface, vertices, graph = "none",
  cost_distance = FALSE)
}
\arguments{
\item{cost_surface}{Cost Surface. Expects Object of class TransitionLayer.}

\item{vertices}{Location from which the Least Cost Paths are calculated. Expects Object of class SpatialPoints}

\item{graph}{graph type used to select which vertices to use. Current implementation accepts 'delauney' (default) or 'none'.}

\item{cost_distance}{computes total cost from each location to destination within the network. Results added to each Least Cost Path within the network. Default is FALSE.}
}
\description{
Calculates a Least Cost Path network which are often, but not exclusively, used in archaeological research.
}
\details{
The function computes the Least Cost Path network. It uses the cost surface(s) generated using the create_slope_cs, create_traversal_cs and or create_openness_cs functions. The function takes a Cost Surface ('TransitionLayer' class) and vertices features ('SpatialPoints' class) for the locations.
}
\author{
Joseph Lewis
}
