% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_openness_cs.R
\name{create_openness_cs}
\alias{create_openness_cs}
\title{create_openness_cs}
\usage{
create_openness_cs(dem, kernel = 3)
}
\arguments{
\item{dem}{Digital Elevation Model. Expects Object of class RasterLayer}

\item{kernel}{Kernel size to calculate openness. Larger kernel size increases the distance that the openness is calculated to. Default kernel size is is 3.}
}
\description{
Creates openness Cost Surface to be used in Least Cost Path calculation.
}
\details{
The create_openness_cs function computes a cost surface based on degree in which a location is sheltered or open to vision
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
create_openness_cs(r, kernel = 3)
}
\author{
Joseph Lewis
}
