% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_attraction.R
\name{feature_attraction}
\alias{feature_attraction}
\title{feature_attraction}
\usage{
feature_attraction(dem, locations, viewshed = NULL, decay = "linear",
  decay_rate = c(5, 500), suffix = "")
}
\arguments{
\item{dem}{Digital Elevation Model. Expects Object of class RasterLayer}

\item{locations}{Location from which the Least Cost Path is calculated. Expects Object of class SpatialPoints or data.frame}

\item{viewshed}{Cost Function to be used in the Least Cost Path calculation. Current implementation computes LCPs using Tobler's Hiking function, Marquez-Perez et al. Modified Hiking function, Herzog's wheeled transport function, and Herzog's sixth degree polynomial function. Default parameter value is is 'all'. See Details for more.}

\item{decay}{If FALSE (default) then Least Cost Paths computed from origin to destination and destination to origin. This is to reflect the ansitropy of Tobler's and Marquez-Perez's cost functions. If TRUE then Least Cost Paths computed from origin to destination only. see Details for more.}

\item{decay_rate}{Number of directions used in the Least Cost Path calculation. \href{https://www.ncbi.nlm.nih.gov/pubmed/17892887}{Huber and Church (1985)} for methodological considerations when considering number of neighbours. Expected input values are 4, 8, 16. Default is 16.}

\item{suffix}{Text to add to end of file name. Useful when calculating least cost paths with different parameters.}
}
\description{
Calculates landscape feature attraction cost RasterLayer
}
\details{
The function computes a landscape feature attraction cost RasterLayer. It implements the method proposed by Llobera (xx) and uses the linear decay rate to decrease the attraction as distance increases from the feature. The function takes a Digital Elevation Model ('RasterLayer' class) and the linear decay rate.
}
\author{
Joseph Lewis
}
