% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_isotropy.R
\name{force_isotropy}
\alias{force_isotropy}
\title{Coerce an anisotropic cost surface to an isotropic cost surface}
\usage{
force_isotropy(x)
}
\description{
Coerce an anisotropic cost surface to an isotropic cost surface
}
