% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{get coordinates from a variety of different object classes}
\usage{
get_coordinates(x)
}
\arguments{
\item{x}{coordinates. \code{sf} 'POINT' or 'MULTIPOINT', \code{SpatVector}, \code{data.frame} or \code{matrix} containing the locations coordinates}
}
\value{
\code{matrix} matrix of coordinates
}
\description{
get coordinates from a variety of different object classes
}
\author{
Joseph Lewis
}
