% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_origin_offset.R
\name{get_origin_offset}
\alias{get_origin_offset}
\title{Calculates an offset value from a given target cell number}
\usage{
get_origin_offset(
  target_cell,
  target_cell_rn,
  cs_ncol,
  num_neigh_cells_from_target
)
}
\arguments{
\item{target_cell}{\code{integer} Cell number from parent raster}

\item{target_cell_rn}{\code{integer} Row number of target cell from parent raster}

\item{cs_ncol}{\code{integer} Number of columns in parent raster}

\item{num_neigh_cells_from_target}{\code{integer} The number of cells between the target cell and the edge of the neighbourhood. For a neighbourhood with dimensions of 5*5, this would equal 2 (i.e. \code{floor(neigh_dim/2)}).}
}
\value{
\code{integer} Offset value to be summed with parent cell values to transform them into values relative to the neighbourhood size.
}
\description{
Calculates an offset value from a given target cell number so that the target cell and its neighbouring cells can be transformed to values relative to the neighbourhood size. Not be called by the end user.
}
\author{
Barney Harris
}
