% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_moves.R
\name{get_reference_moves}
\alias{get_reference_moves}
\title{Generates reference tables of moves available and adjacent cell numbers}
\usage{
get_reference_moves(cs_dem, neighbours, plot_move_paths = F)
}
\arguments{
\item{cs_dem}{\code{spatRaster}}

\item{neighbours}{\code{numeric} value. Number of directions used in the conductance matrix calculation. Expected numeric values are 4, 8, 16, 32, 48. 16 (default)}
}
\value{
\code{list} containing `reference_tables` of class \code{data.table} containing all possible moves from central given as parent cell number \code{cell} and relative cell number \code{rel_next_node}, including interstitial cells \code{rel_next_node_path}, according to the cost-surface \code{neighbourhood} value and supplied \code{cs_dem} dimensions, and \code{vector} of cells immediately adajacent to the central cell. `distance_multipliers` contains the euclidean distance of each move and a multiplier normalized to the distance of a straight move to an adjacent cell. Not be called by the end user.
}
\description{
Generates tables containing all possible moves available to the least-cost path routing algorithm, including interstitial cells, from the central cell, according to the supplied neighborhood value. Move cell numbers are given in values relative to the neighbourhood window value and source \code{cs_dem}.
}
\author{
Barney Harris
}
