% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cs.R
\name{modify_cs}
\alias{modify_cs}
\title{Adjust cost values of a pre-calculated cost surface, using a `SpatVector` object.}
\usage{
modify_cs(
  cs,
  y,
  cost_moving_to = NULL,
  cost_moving_from = NULL,
  cost_moving_between = NULL
)
}
\arguments{
\item{cs}{\code{conductanceMatrix}}

\item{y}{\code{Spat*} object.}

\item{cost_moving_to}{\code{numeric}. Cost of moving from a non-spatially overlapping pixel onto a spatially overlapping pixel. Can be a function where \code{x} is the 'from' node and \code{y} is the 'to' node.}

\item{cost_moving_from}{\code{numeric}. Cost of moving from a spatially overlapping pixel onto a non-spatially overlapping pixel. Can be a function where \code{x} is the 'from' node and \code{y} is the 'to' node.}

\item{cost_moving_between}{\code{numeric}. Cost of moving between spatially overlapping pixels. Can be a function where \code{x} is the 'from' node and \code{y} is the 'to' node.}
}
\value{
returns object of class \code{conductanceMatrixProfile}
}
\description{
Updates nodes within the cppRouting graph, based on whether they spatially overlap with a supplied spatial object. New costs can be supplied when moving from, to and between overlapping nodes.
}
\examples{


r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)
 
cs_profile <- profile_cs(cs = slope_cs)

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

locs <- sf::st_sf(geometry = sf::st_sfc(
  sf::st_point(c(839769, 4199443)),
  sf::st_point(c(1038608, 4100024)),
  sf::st_point(c(907695, 4145478)),
  crs = terra::crs(r)))

ln <- sf::st_combine(locs[1:2,]) \%>\% 
  st_cast('LINESTRING')
y <- vect(ln)

cs_mod <- modify_cs(cs = cs, y = vect(ln), cost_moving_to=200, cost_moving_from=200, cost_moving_between=200)

lcp <- create_lcp(cs = cs_mod, origin = locs, destination = locs, permute = TRUE, plot_paths = FALSE)
}
\author{
Barney Harris
}
