% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhood.R
\name{neighbourhood}
\alias{neighbourhood}
\title{neighbourhood matrix to represent adjacent cells}
\usage{
neighbourhood(neighbours)
}
\arguments{
\item{neighbours}{\code{numeric} or \code{matrix} Expected numeric values are 4, 8, 16, 32, or 48. A user-supplied matrix of 0 and 1s can be supplied. Here, 0 denotes non-adjaency and 1 denotes adjacency}
}
\value{
\code{matrix} neighbourhood matrix
}
\description{
neighbourhood matrix to represent adjacent cells
}
\author{
Joseph Lewis
}
