% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\docType{data}
\name{neighbours_32}
\alias{neighbours_32}
\title{32 Neighbourhood matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 7 columns.
}
\usage{
neighbours_32
}
\description{
32 neighbourhood matrix based on [Kovanen and Sarjakoski (2015)](https://doi.org/10.1145/2803172)
}
\author{
Joseph Lewis
}
\keyword{datasets}
