% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\docType{data}
\name{neighbours_48}
\alias{neighbours_48}
\title{48 Neighbourhood matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
neighbours_48
}
\description{
48 neighbourhood matrix based on [Kovanen and Sarjakoski (2015)](https://doi.org/10.1145/2803172)
}
\author{
Joseph Lewis
}
\keyword{datasets}
