% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slope_cs.R
\name{plot.conductanceMatrix}
\alias{plot.conductanceMatrix}
\title{plot conductanceMatrixWithGraph}
\usage{
\method{plot}{conductanceMatrix}(x, ...)
}
\arguments{
\item{x}{\code{conductanceMatrixWithGraph}}

\item{...}{arguments passed to \code{terra::plot}}
}
\description{
plot conductanceMatrixWithGraph for visualisation. Conductivity values are the mean conductivity for each cell
}
