% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cf.R
\name{plot_cf}
\alias{plot_cf}
\title{plot a conductance function}
\usage{
plot_cf(
  cost_function,
  crit_slope = NULL,
  percentile = NULL,
  rescale = TRUE,
  title = NULL
)
}
\arguments{
\item{cost_function}{\code{character} or \code{function}. Cost function applied to slope values. See details for implemented cost functions. tobler (default)}

\item{crit_slope}{\code{numeric} value. Critical Slope (in percentage) is 'the transition where switchbacks become more effective than direct uphill or downhill paths'. Cost of climbing the critical slope is twice as high as those for moving on flat terrain and is used for estimating the cost of using wheeled vehicles. Default value is 12, which is the postulated maximum gradient traversable by ancient transport (Verhagen and Jeneson, 2012). Critical slope only used in 'wheeled transport' cost function}

\item{percentile}{\code{numeric} value. Travel rate percentile only used in 'campbell 2019' cost_function. Expected numeric values are 0.01, 0.05, 0.10, 0.15, 0.20, 0.25, 0.30, 0.35, 0.40, 0.45, 0.50, 0.55, 0.60, 0.65, 0.70, 0.75, 0.80, 0.85, 0.90, 0.95, 0.99. 0.5 (default)}

\item{rescale}{\code{logical} if TRUE conductance values scaled to between 0 and 1. This can be useful when comparing multiple cost functions}

\item{title}{\code{character} title to be included in plot. NULL (default)}
}
\value{
\code{plot} showing the conductance (inverse of cost) when traversing slope
}
\description{
create a plot showing the conductance (inverse of cost) when traversing slope
}
\details{
The following cost functions have been implemented however users may also supply their own cost functions (see examples):

"tobler", "tobler offpath", "davey", 'rees', "irmischer-clarke male", "irmischer-clarke offpath male", "irmischer-clarke female", "irmischer-clarke offpath female", "modified tobler", 'garmy', 'Kondo-Saino', "wheeled transport", "herzog", "llobera-sluckin", 'naismith', 'minetti', 'campbell', "campbell 2019"
}
\examples{

plot_cf(cost_function = "tobler")
plot_cf(cost_function = function(x) { ((6 * exp(-3.5 * abs(x + 0.05))) * 0.6) / 3.6})

plot_cf(cost_function = "wheeled transport", crit_slope = 8)
plot_cf(cost_function = "campbell 2019", percentile = 0.5)
}
\author{
Joseph Lewis
}
