% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_cs_custom.R
\name{profile_cs_custom}
\alias{profile_cs_custom}
\title{Profile the memory requirements of a cost surface (with origins / destinations) to estimate optimum parameters. Currently under development. Not exported.}
\usage{
profile_cs_custom(
  cs,
  origin,
  destination,
  route_profile_targets = c(100, 225, 625, 1024, 2116),
  iterations = 3
)
}
\arguments{
\item{cs}{\code{conductanceMatrix}}

\item{cost_distance}{\code{logical} if TRUE computes total accumulated cost from origin to the destinations. FALSE (default)}
}
\value{
\code{sf} returns object of class \code{conductanceMatrixProfile}
}
\description{
Profiles the memory resources and processing time required to calculate least-cost paths using a given cost surface and origins / destinations.
}
\examples{


r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

cs_profile <- profile_cs(cs = slope_cs)
}
\author{
Barney Harris
}
