% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_values.R
\name{update_values}
\alias{update_values}
\title{update values in conductanceMatrix}
\usage{
update_values(x, sf, FUN)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{sf}{\code{sf}}

\item{FUN}{\code{function}}
}
\value{
\code{conductanceMatrix}
}
\description{
Update values in conductanceMatrix based on the supplied sf object
}
\details{
the updated conductanceMatrix is produced by assessing which areas of the conductanceMatrix coincide with the supplied sf object. The values within the areas that coincide with the sf object are manipulated based on the supplied function
}
