% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_path_matrix.R
\name{wide_path_matrix}
\alias{wide_path_matrix}
\title{Create a wide path matrix}
\usage{
wide_path_matrix(ncells)
}
\arguments{
\item{ncells}{\code{numeric} value. Dimension of wide path matrix. Note that the value refers to the number of cells and not distance}
}
\value{
\code{matrix} wide path matrix used when calculating wide path least cost paths via \code{\link{create_wide_lcp}}
}
\description{
Creates a wide path matrix to be used when calculating wide path least cost paths. This function will return an odd-dimension matrix approximating the shape of an octogan. The centre cell of the matrix has a value of 0 and represents the focal cell. See \code{\link[raster]{focal}}, \code{\link[raster]{focalWeight}} and \code{\link[raster]{adjacent}} for more information.
}
\examples{
w <- wide_path_matrix(9)
}
\author{
Joseph Lewis
}
