% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dem_error.R
\name{add_dem_error}
\alias{add_dem_error}
\title{Incorporate vertical error into a Digital Elevation Model}
\usage{
add_dem_error(dem, rmse)
}
\arguments{
\item{dem}{\code{spatRaster}}

\item{rmse}{\code{numeric}. Vertical Root Mean Square Error of the Digital Elevation Model}
}
\description{
Incorporate vertical error into a Digital Elevation Model
}
\details{
Digital Elevation Models (DEMs) are representations of the earth's surface and are subject to error (Wechsler, 1999).

The add_dem_error function incorporates vertical error into the supplied DEM by assuming that the error for each cell follows a gaussian (normal) distribution around the measured elevation value and the global Root Mean Square Error (RMSE) estimating the local error variance around this values (Fisher and Tate, 2006). Addition of spatial autocorrelation applied by using a mean-window filter based on a window size (Wechsler and Kroll, 2006).

Examples of RMSE for various datasets:

Shuttle Radar Topography Mission (SRTM) has a RMSE of 9.73m

Advanced Spaceborne Thermal Emission and Reflection Radiometer (ASTER) Global Digital Elevation Model (GDEM) has a RMSE of 10.20m

Ordnance Survey OS Terrain 5 has a maximum RMSE of 2.5m

Ordnance Survey OS Terrain 50 has a maximum RMSE of 4m
}
\author{
Joseph Lewis
}
