% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{calculates distance between adjacent cells}
\usage{
calculate_distance(x, adj)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{adj}{\code{matrix} of adjacent cells}
}
\value{
\code{matrix} euclidean distances between adjacent cells
}
\description{
calculates distance between adjacent cells
}
\author{
Joseph Lewis
}
