% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locations.R
\name{check_locations}
\alias{check_locations}
\title{check supplied locations}
\usage{
check_locations(x, locations)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{locations}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}
}
\value{
\code{message}
}
\description{
checks that locations can be reached when calculating least-cost paths
}
\details{
Using the supplied conductanceMatrix, the function checks whether:
(1) the supplied locations are traversable from at least one adjacent cell
(2) the supplied locations are within the extent of the supplied conductanceMatrix
}
\examples{

r <- terra::rast(system.file("ex/test.grd", package="terra"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler")

locs <- sf::st_sf(geometry = sf::st_sfc(
sf::st_point(c(178743.9, 329740.1)),
sf::st_point(c(180097, 330248.4)),
sf::st_point(c(180714.8, 333283)),
sf::st_point(c(182402.6, 331717.2)),
crs = terra::crs(r)))

check_locations(x = slope_cs, locations = locs)
}
\author{
Joseph Lewis
}
