% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locations.R
\name{check_locations}
\alias{check_locations}
\title{Check locations}
\usage{
check_locations(cost_surface, locations)
}
\arguments{
\item{cost_surface}{\code{TransitionLayer} (gdistance package). Cost surface to be used when checking whether supplied locations are traversable from at least one adjcacent cell}

\item{locations}{\code{SpatialPoints*} (sp package) locations to check}
}
\value{
\code{numeric vector} of location indexes that are not traversable from at least one adjacent cell
}
\description{
Checks that locations can be reached when calculating least cost paths
}
\details{
Using the supplied cost surface, the function checks whether the cell(s) of supplied locations are traversable from at least one adjacent cell. If the cell(s) of the supplied location are not traversable from at least one adjacent cell then a calculated least cost path cannot traverse to that location.
}
\author{
Joseph Lewis
}
