% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_accum_cost.R
\name{create_accum_cost}
\alias{create_accum_cost}
\title{creates an accumulated cost surface}
\usage{
create_accum_cost(x, origin, rescale = FALSE)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{origin}{\code{sf} 'POINT' or 'MULTIPOINT', \code{SpatVector}, \code{data.frame} or \code{matrix} containing the origin coordinates. Only the first row of the supplied object is used as the origin.}

\item{rescale}{\code{logical}. if TRUE, values scaled to between 0 and 1. FALSE (default)}
}
\value{
\code{SpatRaster}
}
\description{
Creates an accumulated cost surfaces from an origin
}
\examples{

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

loc <- sf::st_sf(geometry = sf::st_sfc(sf::st_point(c(839769, 4199443)),crs = terra::crs(r)))

cc <- create_accum_cost(x = slope_cs, origin = loc, rescale = TRUE)
}
\author{
Joseph Lewis
}
