% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_banded_lcps.R
\name{create_banded_lcps}
\alias{create_banded_lcps}
\title{create_banded_lcps}
\usage{
create_banded_lcps(cost_surface, location, min_distance, max_distance,
  radial_points, cost_distance = FALSE, parallel = FALSE)
}
\arguments{
\item{cost_surface}{\code{TransitionLayer} object (gdistance package). Cost surface to be used in Least Cost Path calculation}

\item{location}{\code{SpatialPoints}. Location from which the Least Cost Paths are calculated. Only the first cell is taken into account}

\item{min_distance}{minimum distance from centre location}

\item{max_distance}{maximum distance from centre location}

\item{radial_points}{Number of random locations around centre location within distances}

\item{cost_distance}{if TRUE computes total accumulated cost for each Least Cost Path. Default is FALSE}

\item{parallel}{if TRUE, the Least Cost Paths will be calculated in parallel. Number of Parallel socket clusters is total number of cores available minus 1. Default is FALSE}
}
\value{
SpatialLinesDataFrame object
}
\description{
Calculates Least Cost Paths from random locations within distances
}
\details{
Computes Least Cost Paths from centre location to random locations within distances. This is based on the method proposed by Llobera (2015).
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

locs <- sp::spsample(as(r, 'SpatialPolygons'),n=1,'random')

lcp_network <- create_banded_lcps(cost_surface = final_cost_cs, location = locs, min_distance = 20,
max_distance = 50, radial_points = 10, cost_distance = FALSE, parallel = FALSE)
}
\author{
Joseph Lewis
}
