% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_CCP_lcps.R
\name{create_CCP_lcps}
\alias{create_CCP_lcps}
\title{create_CCP_lcps}
\usage{
create_CCP_lcps(cost_surface, location, distance, radial_points,
  cost_distance = FALSE, parallel = FALSE)
}
\arguments{
\item{cost_surface}{\code{TransitionLayer} object (gdistance package). Cost surface to be used in Least Cost Path calculation}

\item{location}{\code{SpatialPoints}. Location to which the Least Cost Paths are calculated to. Only the first cell is taken into account}

\item{distance}{\code{numeric} value. Distance from centre location to the radial locations}

\item{radial_points}{\code{numeric} value. Number of radial locations around centre location}

\item{cost_distance}{if TRUE computes total accumulated cost for each Least Cost Path. Default is FALSE}

\item{parallel}{if TRUE, the Least Cost Paths will be calculated in parallel. Number of Parallel socket clusters is total number of cores available minus 1. Default is FALSE}
}
\value{
SpatialLinesDataFrame object
}
\description{
Calculates Cumulative Cost Paths from Radial Locations
}
\details{
Computes Least Cost Paths from radial locations of a specified distance to the centre location. This is based on the method proposed by Verhagen (2013).
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

locs <- sp::spsample(as(r, 'SpatialPolygons'),n=1,'random')

lcp_network <- create_CCP_lcps(cost_surface = final_cost_cs, location = locs, distance = 20,
radial_points = 10, cost_distance = FALSE, parallel = FALSE)
}
\author{
Joseph Lewis
}
