% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cost_corridor.R
\name{create_cost_corridor}
\alias{create_cost_corridor}
\title{creates a cost corridor}
\usage{
create_cost_corridor(x, origin, destination, rescale = FALSE)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{origin}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}

\item{destination}{\code{sf} of geometry type 'POINT' or 'MULTIPOINT'}

\item{rescale}{\code{logical}. if TRUE, values scaled to between 0 and 1. FALSE (default)}
}
\value{
\code{SpatRaster}
}
\description{
Combines the accumulated cost surfaces from origin-to-destination and destination-to-origin to identify areas of preferential movement that takes into account both directions of movement
}
\author{
Joseph Lewis
}
