% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_cs.R
\name{create_distance_cs}
\alias{create_distance_cs}
\title{Creates a distance-based cost surface}
\usage{
create_distance_cs(x, neighbours = 16, max_slope = NULL, exaggeration = FALSE)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{neighbours}{\code{numeric} value. Number of directions used in the conductance matrix calculation. Expected numeric values are 4, 8, 16, 32, 48. 16 (default)}

\item{max_slope}{\code{numeric} value. Maximum percentage slope that is traversable. Slope values that are greater than the specified max_slope are given a conductivity value of 0. If cost_function argument is 'campbell 2019' then max_slope is fixed at 30 degrees slope to reflect the maximum slope that the cost function is parametised to. NULL (default)}

\item{exaggeration}{\code{logical}. if TRUE, positive slope values (up-hill movement) multiplied by 1.99 and negative slope values (down-hill movement) multiplied by 2.31}
}
\value{
\code{conductanceMatrix} that numerically expresses the difficulty of moving across slope based on the provided cost function
}
\description{
Creates a cost surface based on the distance between neighbouring cells. Distance corrected for if neighbours value is greater than 4.
}
\examples{

r <- terra::rast(system.file("ex/test.grd", package="terra"))

slope_cs <- create_distance_cs(x = r)

}
\author{
Joseph Lewis
}
