% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_FETE_lcps.R
\name{create_FETE_lcps}
\alias{create_FETE_lcps}
\title{Calculate least cost paths from each location to all other locations.}
\usage{
create_FETE_lcps(
  cost_surface,
  locations,
  cost_distance = FALSE,
  parallel = FALSE,
  ncores = 1
)
}
\arguments{
\item{cost_surface}{\code{TransitionLayer} (gdistance package). Cost surface to be used in Least Cost Path calculation}

\item{locations}{\code{SpatialPoints*} (sp package). Locations to calculate Least Cost Paths from and to}

\item{cost_distance}{\code{logical}. if TRUE computes total accumulated cost for each Least Cost Path. Default is FALSE}

\item{parallel}{\code{logical}. if TRUE, the Least Cost Paths will be calculated in parallel. Default is FALSE}

\item{ncores}{\code{numeric}. Number of cores used if parallel is TRUE. Default value is 1.}
}
\value{
\code{SpatialLinesDataFrame} (sp package). The resultant object contains least cost paths calculated from each location to all other locations
}
\description{
Calculates least cost paths from each location to all other locations (i.e. From Everywhere To Everywhere (FETE)). This is based on the method proposed by White and Barber (2012).
}
\examples{

r <- raster::raster(nrow=50, ncol=50,  xmn=0, xmx=50, ymn=0, ymx=50,
crs='+proj=utm')

r[] <- stats::runif(1:length(r))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

locs <- sp::spsample(as(raster::extent(r), 'SpatialPolygons'),n=5,'regular')

lcp_network <- create_FETE_lcps(cost_surface = slope_cs, locations = locs,
cost_distance = FALSE, parallel = FALSE)
}
\references{
White, DA. Barber, SB. (2012). Geospatial modeling of pedestrian transportation networks: a case study from precolumbian Oaxaca, Mexico. J Archaeol Sci 39:2684-2696. \doi{10.1016/j.jas.2012.04.017}
}
\author{
Joseph Lewis
}
