% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp.R
\name{create_lcp}
\alias{create_lcp}
\title{Calculate Least Cost Path from Origin to Destination}
\usage{
create_lcp(x, origin, destination, directional = FALSE, cost_distance = FALSE)
}
\description{
Calculate Least Cost Path from Origin to Destination
}
\author{
Joseph Lewis
}
