% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp.R
\name{create_lcp}
\alias{create_lcp}
\title{create_lcp}
\usage{
create_lcp(cost_surface, origin, destination, directional = FALSE)
}
\arguments{
\item{cost_surface}{Cost Surface. Expects Object of class TransitionLayer.}

\item{origin}{Location from which the Least Cost Path is calculated. Expects Object of class SpatialPoints}

\item{destination}{Location to which the Least Cost Path is calculated. Expects Object of class SpatialPoints}

\item{directional}{If FALSE (default) then Least Cost Paths computed from origin to destination and destination to origin.}
}
\description{
Calculates Least Cost Paths which are often, but not exclusively, used in archaeological research
}
\details{
The function computes the Least Cost Path from an origin location to a destination location. It uses the cost surface generated using the create_slope_cs and/or create_traversal_cs functions. The function takes a Cost Surface ('TransitionLayer' class) and point features ('SpatialPoints' class) for the origin location and destination location.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

loc1 = cbind(2667670, 6479000)
loc1 = sp::SpatialPoints(loc1)

loc2 = cbind(2667800, 6479400)
loc2 = sp::SpatialPoints(loc2)

lcps <- create_lcp(cost_surface = final_cost_cs, origin = loc1,
destination = loc2, directional = FALSE)
}
\author{
Joseph Lewis
}
