% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp_density.R
\name{create_lcp_density}
\alias{create_lcp_density}
\title{create_lcp_density}
\usage{
create_lcp_density(lcps, raster, rescale = FALSE)
}
\arguments{
\item{lcps}{\code{SpatialLines} or \code{SpatialLinesDataFrame}. Least Cost Paths}

\item{raster}{\code{RasterLayer} (raster package). This is used to derive the resolution, extent, and spatial reference system to be used when calculating the cumulative least cost path raster}

\item{rescale}{if TRUE raster values scaled to between 0 and 1. Default is FALSE}
}
\value{
RasterLayer object
}
\description{
Creates a cumulative least cost path raster
}
\details{
Cumulatively combines Least Cost Paths in order to identify routes of preferential movement within the landscape.
}
\examples{

#r <- raster::raster(nrow=50, ncol=50,  xmn=0, xmx=50, ymn=0, ymx=50, crs='+proj=utm')

#r[] <- stats::runif(1:length(r))

#slope_cs <- create_slope_cs(r, cost_function = 'tobler')

#locs <- sp::spsample(as(r, 'SpatialPolygons'),n=10,'regular')

#lcp_network <- create_FETE_lcps(cost_surface = slope_cs, locations = locs,
#cost_distance = FALSE, parallel = FALSE)

#cumulative_lcps <- create_lcp_density(lcps = lcp_network, raster = r, rescale = FALSE)
}
\author{
Joseph Lewis
}
