% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp_density.R
\name{create_lcp_density}
\alias{create_lcp_density}
\title{creates a cumulative least-cost path layer}
\usage{
create_lcp_density(x, lcps, rescale = FALSE)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{lcps}{\code{sf}}

\item{rescale}{\code{logical}. if TRUE, values scaled to between 0 and 1. FALSE (default)}
}
\value{
\code{SpatRaster}
}
\description{
Cumulatively combines least-cost paths to idenify routes of preferential movement
}
\author{
Joseph Lewis
}
