% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp_density.R
\name{create_lcp_density}
\alias{create_lcp_density}
\title{create_lcp_density}
\usage{
create_lcp_density(lcps, raster, rescale = FALSE)
}
\arguments{
\item{lcps}{\code{SpatialLines} or \code{SpatialLinesDataFrame}. Least Cost Paths}

\item{raster}{\code{RasterLayer} (raster package). This is used to derive the resolution, extent, and spatial reference system to be used when calculating the cumulative least cost path raster}

\item{rescale}{if TRUE raster values scaled to between 0 and 1. Default is FALSE}
}
\value{
RasterLayer object
}
\description{
Creates a cumulative least cost path raster
}
\details{
Cumulatively combines Least Cost Paths in order to identify routes of preferential movement within the landscape.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

locs <- sp::spsample(as(r, 'SpatialPolygons'),n=10,'regular')

lcp_network <- create_FETE_lcps(cost_surface = final_cost_cs, locations = locs,
cost_distance = FALSE, parallel = FALSE)

cumulative_lcps <- create_lcp_density(lcps = lcp_network, raster = r, rescale = FALSE)
}
\author{
Joseph Lewis
}
