% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lcp_network.R
\name{create_lcp_network}
\alias{create_lcp_network}
\title{create_lcp_network}
\usage{
create_lcp_network(cost_surface, locations, cost_distance = FALSE,
  parallel = FALSE)
}
\arguments{
\item{cost_surface}{Cost Surface. Expects Object of class TransitionLayer.}

\item{locations}{from which the Least Cost Paths are calculated. Expects Object of class SpatialPoints}

\item{cost_distance}{computes total cost from each location to destination within the network. Results added to each Least Cost Path within the network. Default is FALSE.}

\item{parallel}{if TRUE, the LCPs will be calculated in parallel. Number of Parallel socket clusters is total number of cores minus 1. Default is FALSE}
}
\description{
Calculates a Least Cost Path network which are often, but not exclusively, used in archaeological research.
}
\details{
The function computes the Least Cost Path network. It uses a cost surface generated using the functions within the leastcostpath library. The create_lcp_network function expects a Cost Surface ('TransitionLayer' class) and SpatialPoints ('SpatialPoints' class) for the locations.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))

slope_cs <- create_slope_cs(r, cost_function = 'tobler')

traverse_cs <- create_traversal_cs(r, neighbours = 16)

final_cost_cs <- slope_cs * traverse_cs

locs <- sp::spsample(as(r, 'SpatialPolygons'),n=10,'regular')

lcp_network <- create_lcp_network(cost_surface = final_cost_cs, locations = locs,
cost_distance = FALSE, parallel = FALSE)
}
\author{
Joseph Lewis
}
