% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slope_cs.R
\name{create_slope_cs}
\alias{create_slope_cs}
\title{Creates a slope-based cost surface}
\usage{
create_slope_cs(
  x,
  cost_function = "tobler",
  neighbours = 16,
  crit_slope = 12,
  max_slope = NULL,
  percentile = 0.5,
  exaggeration = FALSE
)
}
\arguments{
\item{x}{\code{SpatRaster}}

\item{cost_function}{\code{character} or \code{function}. Cost function applied to slope values. See details for implemented cost functions. tobler (default)}

\item{neighbours}{\code{numeric} value. Number of directions used in the conductance matrix calculation. Expected numeric values are 4, 8, 16, 32, 48. 16 (default)}

\item{crit_slope}{\code{numeric} value. Critical Slope (in percentage) is 'the transition where switchbacks become more effective than direct uphill or downhill paths'. Cost of climbing the critical slope is twice as high as those for moving on flat terrain and is used for estimating the cost of using wheeled vehicles. Default value is 12, which is the postulated maximum gradient traversable by ancient transport (Verhagen and Jeneson, 2012). Critical slope only used in 'wheeled transport' cost function}

\item{max_slope}{\code{numeric} value. Maximum percentage slope that is traversable. Slope values that are greater than the specified max_slope are given a conductivity value of 0. If cost_function argument is 'campbell 2019' then max_slope is fixed at 30 degrees slope to reflect the maximum slope that the cost function is parametised to. NULL (default)}

\item{percentile}{\code{numeric} value. Travel rate percentile only used in 'campbell 2019' cost_function. Expected numeric values are 0.01, 0.05, 0.10, 0.15, 0.20, 0.25, 0.30, 0.35, 0.40, 0.45, 0.50, 0.55, 0.60, 0.65, 0.70, 0.75, 0.80, 0.85, 0.90, 0.95, 0.99. 0.5 (default)}

\item{exaggeration}{\code{logical}. if TRUE, positive slope values (up-hill movement) multiplied by 1.99 and negative slope values (down-hill movement) multiplied by 2.31}
}
\value{
\code{conductanceMatrix} that numerically expresses the difficulty of moving across slope based on the provided cost function
}
\description{
Creates a cost surface based on the difficulty of moving up and down slope. This function provides multiple isotropic and anisotropic cost functions that estimate the 'cost' of human movement across a landscape.
 
The supplied 'spatRaster' object must have a projected CRS
}
\details{
The following cost functions have been implemented however users may also supply their own cost functions (see examples):

"tobler", "tobler offpath", "davey", 'rees', "irmischer-clarke male", "irmischer-clarke offpath male", "irmischer-clarke female", "irmischer-clarke offpath female", "modified tobler", 'garmy', 'Kondo-Saino', "wheeled transport", "herzog", "llobera-sluckin", 'naismith', 'minetti', 'campbell', "campbell 2019"
}
\examples{

r <- terra::rast(system.file("ex/test.grd", package="terra"))

slope_cs <- create_slope_cs(x = r, cost_function = "tobler")
slope_cs2 <- create_slope_cs(x = r, 
cost_function = function(x) {(6 * exp(-3.5 * abs(x + 0.05))) / 3.6})

}
\author{
Joseph Lewis
}
