% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traversal_cs.R
\name{create_traversal_cs}
\alias{create_traversal_cs}
\title{create_traversal_cs}
\usage{
create_traversal_cs(dem, neighbours = 16)
}
\arguments{
\item{dem}{Digital Elevation Model. \code{RasterLayer} (raster package)}

\item{neighbours}{Number of directions used in the Least Cost Path calculation. See Huber and Church (1985) for methodological considerations when choosing number of neighbours. Expected values are 4, 8, or 16. Default is 16}
}
\value{
Transition object
}
\description{
Creates a Traversal across slope Cost Surface
}
\details{
Creates a cost surface based on the difficulty of traversing **(across)** slope. Difficulty of traversal is based on the figure given in Bell and Lock (2000). Traversal across slope accounts for movement directly perpendicular across slope being easier than movement diagonally up/down slope.
}
\examples{
r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
traversal_cs <- create_traversal_cs(r, neighbours = 16)
}
\author{
Joseph Lewis
}
