% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_attraction.R
\name{feature_attraction}
\alias{feature_attraction}
\title{feature_attraction}
\usage{
feature_attraction(dem, locations, viewshed = NULL, decay = "linear",
  decay_rate = c(5, 500), suffix = "")
}
\arguments{
\item{dem}{Digital Elevation Model. Expects Object of class RasterLayer}

\item{locations}{Locations of landscape features. Expects Object of class SpatialPoints or data.frame}

\item{viewshed}{if viewshed is supplied with a RasterLayer then landscape feature attraction values are limited to only visible areas.}

\item{decay}{Default is "linear". Exponential to be implemented soon.}

\item{decay_rate}{Expects vector of two values denoting the Rank of the landscape feature and the maximum attraction value. Default is c(5, 500). scale in metres (m).}

\item{suffix}{Text to add to end of file name. Useful when calculating least cost paths with different parameters.}
}
\description{
Calculates landscape feature attraction cost RasterLayer
}
\details{
The function computes a landscape feature attraction cost RasterLayer. It implements the method proposed by Llobera (2000) and uses the linear decay rate to decrease the attraction as distance increases from the feature. The function takes a Digital Elevation Model ('RasterLayer' class) and the linear decay rate.
}
\author{
Joseph Lewis
}
