% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_isotropy.R
\name{force_isotropy}
\alias{force_isotropy}
\title{Coerce an anisotropic cost surface to an isotropic cost surface}
\usage{
force_isotropy(x)
}
\arguments{
\item{x}{\code{conductanceMatrix}}
}
\value{
\code{conductanceMatrix} 

r <- terra::rast(system.file("extdata/SICILY_1000m.tif", package="leastcostpath"))

slope_cs_aniso <- create_slope_cs(x = r, cost_function = "tobler", neighbours = 4)

slope_cs_iso <- force_isotropy(slope_cs_aniso)
}
\description{
Averages conductance values from-to adjacent cells
}
