% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\docType{data}
\name{neighbours_32}
\alias{neighbours_32}
\title{32 Neighbourhood matrices based on Kovanen and Sarjakoski (2015)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 7 columns.
}
\usage{
neighbours_32
}
\description{
see leastcostpath::neighbours_32 for layout
}
\references{
Kovanen, J., Sarjakoski, T. (2015). Tilewise Accumulated Cost Surface Computation with Graphics Processing Units. ACM Transactions on Spatial Algorithms and Systems 1, 1-27. \url{https://doi.org/10.1145/2803172}
}
\author{
Joseph Lewis
}
\keyword{datasets}
