% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\docType{data}
\name{neighbours_48}
\alias{neighbours_48}
\title{48 Neighbourhood matrices based on Kovanen and Sarjakoski (2015)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
neighbours_48
}
\description{
see leastcostpath::neighbours_48 for layout
}
\references{
Kovanen, J., Sarjakoski, T. (2015). Tilewise Accumulated Cost Surface Computation with Graphics Processing Units. ACM Transactions on Spatial Algorithms and Systems  1, 1-27. \doi{10.1145/2803172}
}
\author{
Joseph Lewis
}
\keyword{datasets}
