% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDI_validation.R
\name{PDI_validation}
\alias{PDI_validation}
\title{Calculate Path Deviation Index}
\usage{
PDI_validation(lcp, comparison, normalise = FALSE)
}
\arguments{
\item{lcp}{\code{SpatialLines*} (sp package). Least Cost Path to assess the accuracy of. Expects object of class SpatialLines.}

\item{comparison}{\code{SpatialLines*} to validate the Least Cost Path against.}

\item{normalise}{\code{logical}. If TRUE, PDI is normalised to allow for comparisons when PDIs are calculated for paths with different origin and destination points. Default is FALSE.}
}
\value{
\code{numeric value} (base package). The average distance between two paths and is calculated as the area between paths divided by the shortest path (straight line distance) between the start and end point.
}
\description{
Calculates the Path Deviation Index of a Least Cost Path and a comparison SPatialLines using the method proposed by Jan et al. (1999).
}
\examples{
x1 <- c(1,5,4,50)
y1 <- c(1,3,4,50)
line1 <- sp::SpatialLines(list(sp::Lines(sp::Line(cbind(x1,y1)), ID='a')))
x2 <- c(1,5,5,50)
y2 <- c(1,4,6,50)
line2 <- sp::SpatialLines(list(sp::Lines(sp::Line(cbind(x2,y2)), ID='b')))

val_lcp <- PDI_validation(lcp = line1, line2)
}
\references{
Jan, O., Horowitz, A.J., Peng, Z,R. 1999. Using GPS data to understand variations in path choice. Paper presented at the 78th meeting of the Transportation Research Board, Washington.
}
\author{
Joseph Lewis
}
